requirements.txt
mujoco-py<1.50.2,>=1.50.1
seaborn==0.9.0
pandas==1.0.3
cloudpickle==1.3.0
gym==0.10.11
scipy==1.1.0
tensorflow==1.7.1 
numpy==1.16.1
opencv_contrib_python==4.1.0.25
matplotlib==2.5.5
h5py==2.7.1
baselines==0.1.5
stable-baselines==2.6.0
mpi4py==3.0.3
cma==3.0.3
pyyaml
sklearn
pytablewriter